Update absorption coefficient and frequency units
Update all private functions such that input and output are all in SI units
Update all public functions such that input and output are in conventional units i.e. absorption coefficient in cm-1, frequency in THz

Exceptions:
`_transform`: frequency input and output are both in THz based on the consideration of computational efficiency
`_all_optical_constants`: input are in SI units and output are in conventional units. This is unique since the def takes in values from private functions and convert it to a form suitable for output in public functions